api interfaces done as of now

+++ app/src/main/java/com/sandeveloper/jsscolab/domain/Modules/Auth/ForgetPasswordRequest.kt
@@ -1,7 +1,7 @@
 package com.sandeveloper.jsscolab.domain.Modules.Auth
 
 data class ForgetPasswordRequest (
-    val phone:Int,
+    val phone:Long,
     val password:String,
     val temp_token:String
 )
