+++ app/src/main/java/com/sandeveloper/jsscolab/domain/Modules/Auth/OtpRequest.kt
@@ -1,6 +1,6 @@
 package com.sandeveloper.jsscolab.domain.Modules.Auth
 
 data class OtpRequest(
-    val phone:Int,
+    val phone:Long,
     val action:String
 )
